Create HumareadAbleTimeToEpochTimeConverter
It is a simple C program that converts human-readable time strings to epoch time format. It provides a method convertToEpochTime which takes a human-readable time string as input and returns the corresponding epoch time. This program can be useful for developers working with date and time manipulation in C applications. Very useful in Loadrunner